reorganized and wrote script to update num_int_pts of a db record
