+++ lmfdb/elliptic_curves/templates/ec-index.html
@@ -204,7 +204,7 @@ A <a href={{url_for('.random_curve')}}>random elliptic curve</a> from the databa
     Number of {{ KNOWL('ec.q.integral_points',title="integral points") }}
     </td>
     <td>
-    <input type='text' name='num_int_pts' example="2" size=10 />
+    <input type='text' name='num_int_pts' example="2" size=10>
     </td>
     <td>
     <span class="formexample"> e.g. 2 or 4-15</span>
