+++ lmfdb/elliptic_curves/templates/ec-search-results.html
@@ -88,6 +88,7 @@
 
 <tr>
 <td align=left colspan=2>{{KNOWL('ec.q.reduction_type', title='Bad primes')}}</td>
+<td>Number of {{ KNOWL('ec.q.integral_points',title="integral points") }}</td>
 </tr>
 
 <tr>
@@ -108,6 +109,9 @@
 {% endif %}
   </select>
 <input type='text' name='bad_primes' placeholder="2,3" size=10  value={{info.bad_primes}}>
+</td>
+<td>
+<input type='text' name='num_int_pts' example="2" size=10>
 </td>
 </tr>
