+++ scripts/elliptic_curves/import_ec_data.py
@@ -151,7 +151,7 @@ import re, os
 import time
 from sage.all import ZZ, RR, EllipticCurve, prod, Set, magma, prime_range, GF, pari
 from lmfdb.utils import web_latex
-from lmfdb.elliptic_curves.web_ec import make_y_coord
+from lmfdb.elliptic_curves.web_ec import make_y_coord, count_integral_points
 from lmfdb import db
 print "setting curves"
 curves = db.ec_curves
@@ -192,7 +192,6 @@ def split(line):
 
 lmfdb_label_to_label = {}
 label_to_lmfdb_label = {}
-    
 
 def allbsd(line):
     r""" Parses one line from an allbsd file.  Returns the label and a
@@ -286,7 +285,7 @@ def allgens(line):
     N = E.conductor()
     bad_p = N.prime_factors() # will be sorted
     num_bad_p = len(bad_p)
-    
+
     local_data = [{'p': int(ld.prime().gen()),
                            'ord_cond':int(ld.conductor_valuation()),
                            'ord_disc':int(ld.discriminant_valuation()),
@@ -311,14 +310,14 @@ def allgens(line):
         minq_ainvs = Etw.ainvs()
         r = curves.lucky({'jinv':str(E.j_invariant()), 'ainvs':minq_ainvs}, projection=['label','lmfdb_label'])
         min_quad_twist = {'label': r['label'], 'lmfdb_label':r['lmfdb_label'], 'disc':int(Dtw)}
-    
-    #print("computing hash")    
+
+    #print("computing hash")
     #trace_hash = ZZ(magma.TraceHash(E))
     #trace_hash = ZZ(magma.eval("TraceHash(EllipticCurve({}));".format(data[3])))
     trace_hash = TraceHashClass(iso, E)
     #trace_hash = ZZ(0)
-    #print("done")    
-        
+    #print("done")
+
     return label,  {
         'conductor': int(data[0]),
         'iso': iso,
@@ -571,7 +570,7 @@ def allisog(line, lmfdb_order=True):
     data = split(line)
     isomat = data[5]
     isomat = [[int(d) for d in row.split(",")] for row in isomat[2:-2].split("],[")]
-    
+
     # This dict stores the degrees of (cyclic) isogenies from each
     # curve.  We must compute this *before* the optional (but default)
     # reordering of the rows/cols.  The value for key n (=1,2,3,...)
@@ -643,7 +642,7 @@ def fix_isogeny_degrees(C):
     #print("changing isogeny_degrees for {} ({}) from {} to {}".format(C['label'],C['lmfdb_label'],old_isodegs,new_isodegs))
     C['isogeny_degrees'] = new_isodegs
     return C
-    
+
 
 def cmp_label(lab1, lab2):
     from sage.databases.cremona import parse_cremona_label, class_to_int
@@ -674,7 +673,7 @@ def copy_records_to_file(records, fname, id0=0, verbose=True):
         print("Writing records to {} and {}".format(searchfile, extrafile))
 
     # NB since records might be a generator object we only pass through it once.
-    
+
     fs = open(searchfile, 'w')
     fe = open(extrafile, 'w')
     curves._write_header_lines(fs, ["id"]+curves.search_cols)
@@ -691,7 +690,7 @@ def copy_records_to_file(records, fname, id0=0, verbose=True):
     fe.close()
     if verbose:
         print("Wrote {} lines to {} and {}".format(id-id0, searchfile, extrafile))
-    
+
 #
 
 
@@ -823,7 +822,7 @@ def upload_to_db(base_path, min_N, max_N, insert=True, mode='test'):
     else:
         print("Some records have missing keys, no uploading")
         return
-    
+
     if insert:
         if mode=='test':
             print("(not) inserting all data")
@@ -1030,7 +1029,7 @@ def check_database_consistency(table, N1=None, N2=None, iwasawa_bound=150000):
 
     tor_gro_keys = ['tor_gro', 'tor_degs', 'tor_fields']
     tor_gro_bound = 400000
-    
+
     print("key_set has {} keys".format(len(key_set)))
 
     query = {}
@@ -1123,7 +1122,7 @@ def update_stats(recount=True, verbose=True):
         ncu = ec_count({'sha':s**2})
         if verbose and ncu:
             print("{} curves have Sha order {}^2".format(ncu,s))
-    
+
 def update_torsion_growth_stats(verbose=True):
     # torsion growth:
     if verbose:
@@ -1255,8 +1254,8 @@ def fix_quad_twist(c):
             mqt['label'] = ct['label']
             if mqt['lmfdb_label'] == '':
                 mqt['lmfdb_label'] = ct['lmfdb_label']
-    return c    
-        
+    return c
+
 
 # Rewrite function to add optimality and manin constant data.
 #
@@ -1291,42 +1290,18 @@ opt_man_data = {} # to keep pyflakes happy
 #
 # before using the following function for rewriting
 
-def count_integral_points(c):
-    ainvs = c['ainvs']
-    num_int_pts = 0
-    #int_pts_str = ""
-    for x in xcoord_integral_points:
-        y = make_y_coord(ainvs,x)
-        if y == 0:
-            num_int_pts += 1
-            #int_pts_str += "({},{}),".format(x,y)
-        else:
-            num_int_pts += 2
-            #int_pts_str += "({},\\pm{}),".format(x,y)
-        #int_pts_str = int_pts_str[:len(int_pts_str)-1]
-    return num_int_pts
-
 def add_opt_man(c):
     lab = c['label']
     if lab in opt_man_data:
         c.update(opt_man_data[lab])
     else:
         print("No new optimality/Manin data for curve {}".format(lab))
-    m = count_integral_points(c)
-    c['num_int_pts'] = m
-    return
+    return c
 
-def count_integral_points(c):
-    ainvs = c['ainvs']
-    num_int_pts = 0
-    #int_pts_str = ""
-    for x in xcoord_integral_points:
-        y = make_y_coord(ainvs,x)
-        if y == 0:
-            num_int_pts += 1
-        else:
-            num_int_pts += 2
-    return num_int_pts
+def update_num_int_pts(rec):
+    from lmfdb.elliptic_curves.web_ec import make_y_coord, make_integral_points, count_integral_points
+    rec['num_int_pts'] = count_integral_points(rec)
+    return rec
 
 # Sage translation of the Magma function TraceHash(), just for elliptic curves /Q
